Update
The app was not working properly due to changes in iOS over last years. I followed some instructions which can be found under the following links (additions and adjustments in .plist, elimination of Scene Delegate's class and methods):

https://www.hackingwithswift.com/forums/100-days-of-swift/project-25-app-crashes-when-try-to-host-a-session/3944
https://stackoverflow.com/questions/64040617/multipeerconnectivity-does-not-work-as-expected-on-ios-14
https://stackoverflow.com/questions/57467003/opt-out-of-uiscenedelegate-swiftui-on-ios

I also added two more constraints in Interface Builder for the cells to display appropriately.